Implementation ultra-fast! which is O(number or live cells)
I had promised not to touch this code anymore. My fault... :-(

But the previous implementation, using matrices, was O(world dimension)
which means really slow. The implementation in this commit uses
a hash map of positions, which means the "world" can now have infinite
size (well, infinite == +- max_value(int)) and this will have no
influence on impact on the algorithm performance at all!

Rest in peace, game of life. I spent too much time on you.

But it was fun :-)